1. Произведена первая часть рефакторинга, часть логики перенесена из MovieQuizViewController в MovieQuizViewPresenter
