+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,38 @@
+//
+//  MovieQuizPresenter.swift
+//  MovieQuiz
+//
+//  Created by Алишер Дадаметов on 17.05.2023.
+//
+
+import UIKit
+
+final class MovieQuizPresenter {
+    
+    private var currentQuestionIndex: Int = 0
+    let questionsAmount: Int = 10
+    
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionsAmount - 1
+    }
+    
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+            
+        let questionStep = QuizStepViewModel(
+            image: UIImage(data: model.image) ?? UIImage(),
+            question: model.text,
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)"
+        )
+        return questionStep
+        
+    }
+}
+
